requirements.txt
fastapi==0.115.6
imagekitio==4.1.0
passlib==1.7.4
pydantic==2.10.5
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.36
